Consolidate resources into single Management context
