--- lib/mt_kitchen/kitchen.ex
@@ -1,126 +1,0 @@
-defmodule MTKitchen.Kitchen do
-  @moduledoc """
-  The Kitchen context. Handles operations relating to recipe and other resource related content.
-    Generally meant to be read-only for most use-cases.
-  """
-
-  import Ecto.Query, warn: false
-  alias MTKitchen.Repo
-
-  alias MTKitchen.Kitchen.Recipe
-
-  @doc """
-  Returns the list of recipes.
-
-  ## Examples
-
-      iex> list_recipes()
-      [%Recipe{}, ...]
-
-  """
-  def list_recipes do
-    Repo.all(Recipe)
-  end
-
-  @doc """
-  Returns the list of recipes displayed to the public
-  """
-  def directory_recipes do
-    Repo.all(
-      from r in Recipe,
-        where: r.publicly_accessible == true
-    )
-  end
-
-  @doc """
-  Gets a single recipe.
-
-  Raises `Ecto.NoResultsError` if the Recipe does not exist.
-
-  ## Examples
-
-      iex> get_recipe!(123)
-      %Recipe{}
-
-      iex> get_recipe!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_recipe!(id), do: Repo.get!(Recipe, id)
-
-  def get_full_recipe!(id) do
-    Repo.one!(
-      from r in Recipe,
-        where:
-          r.id == ^id or
-            (r.slug == ^id and
-               r.publicly_accessible == true),
-        preload: [steps: [step_ingredients: [:ingredient]]]
-    )
-  end
-
-  @doc """
-  Creates a recipe.
-
-  ## Examples
-
-      iex> create_recipe(%{field: value})
-      {:ok, %Recipe{}}
-
-      iex> create_recipe(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_recipe(attrs \\ %{}) do
-    %Recipe{}
-    |> Recipe.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a recipe.
-
-  ## Examples
-
-      iex> update_recipe(recipe, %{field: new_value})
-      {:ok, %Recipe{}}
-
-      iex> update_recipe(recipe, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_recipe(%Recipe{} = recipe, attrs) do
-    recipe
-    |> Recipe.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a recipe.
-
-  ## Examples
-
-      iex> delete_recipe(recipe)
-      {:ok, %Recipe{}}
-
-      iex> delete_recipe(recipe)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_recipe(%Recipe{} = recipe) do
-    Repo.delete(recipe)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking recipe changes.
-
-  ## Examples
-
-      iex> change_recipe(recipe)
-      %Ecto.Changeset{data: %Recipe{}}
-
-  """
-  def change_recipe(%Recipe{} = recipe, attrs \\ %{}) do
-    Recipe.changeset(recipe, attrs)
-  end
-end
