+++ lib/mt_kitchen_web/live/recipe_live/index.html.heex
@@ -5,7 +5,7 @@
     <ul id="directory_recipes" class="directory-items">
         <%= for recipe <- @recipes do %>
             <li class="directory-item">
-                <span><%= live_redirect recipe.name, to: Routes.recipe_show_path(@socket, :show, recipe) %></span>
+                <span><%= live_redirect recipe.name, to: Routes.recipe_show_path(@socket, :show, recipe.slug) %></span>
             </li>
         <% end %>
     </ul>
