--- test/mt_kitchen/kitchen_test.exs
@@ -1,28 +1,0 @@
-defmodule MTKitchen.KitchenTest do
-  use MTKitchen.DataCase
-
-  #  alias MTKitchen.Kitchen
-
-  #  describe "recipes" do
-  #    #    alias MTKitchen.Kitchen.Recipe
-  #
-  #    #    @invalid_attrs %{}
-  #
-  #    #    alias MTKitchen.ManagementFixtures
-  #
-  #    #    test "list_recipes/0 returns all recipes" do
-  #    #      recipe = ManagementFixtures.recipe_fixture()
-  #    #      assert Kitchen.list_recipes() == [recipe]
-  #    #    end
-  #    #
-  #    #    test "get_recipe!/1 returns the recipe with given id" do
-  #    #      recipe = ManagementFixtures.recipe_fixture()
-  #    #      assert Kitchen.get_recipe!(recipe.id) == recipe
-  #    #    end
-  #    #
-  #    #    test "change_recipe/1 returns a recipe changeset" do
-  #    #      recipe = ManagementFixtures.recipe_fixture()
-  #    #      assert %Ecto.Changeset{} = Kitchen.change_recipe(recipe)
-  #    #    end
-  #  end
-end
